small fixes xD

+++ gsoy_importer/util.py
@@ -1,10 +1,8 @@
 import os
+import shutil
 import tarfile
 import urllib.request
 from datetime import datetime, timedelta, timezone
-
-LAST_RUN_LAST_RUN_FILE_PATH = "/last_run/last_run.txt"
-GSOY_DOWNLOAD_URL = "https://www.ncei.noaa.gov/data/gsoy/archive/gsoy-latest.tar.gz"
 
 
 def should_run():
@@ -15,6 +13,7 @@ def should_run():
      :rtype: bool
      """
     from logging_config import logger
+    from config import LAST_RUN_LAST_RUN_FILE_PATH
 
     if not os.path.exists(LAST_RUN_LAST_RUN_FILE_PATH):
         logger.info('No last_run_file found.')
@@ -53,6 +52,10 @@ def update_last_run():
     :return: None
     :rtype: None
     """
+
+    from influx import write_points_to_db
+    from config import LAST_RUN_LAST_RUN_FILE_PATH
+
     current_time = datetime.now()
 
     last_run_point = [
@@ -68,28 +71,89 @@ def update_last_run():
         }
     ]
 
-    from influx import write_points_to_db
     write_points_to_db(last_run_point)
+
+    os.makedirs(os.path.dirname(LAST_RUN_LAST_RUN_FILE_PATH), exist_ok=True)
 
     with open(LAST_RUN_LAST_RUN_FILE_PATH, "w") as f:
         f.write(current_time.strftime("%Y-%m-%d %H:%M:%S"))
 
 
+def extract_tar_file(file_name):
+    """
+    Extracts the tar file to the 'extracted_data' directory within the target directory.
+
+    :param file_name: Name of the tar file to extract.
+    :type file_name: str
+    :return: None
+    :rtype: None
+    """
+    from config import GSOY_DATA_DIR
+    from logging_config import logger
+
+    extracted_data_dir = os.path.join(GSOY_DATA_DIR, 'extracted_data')
+    os.makedirs(extracted_data_dir, exist_ok=True)
+
+    tar = tarfile.open(file_name, 'r:gz')
+    logger.info('Extracting tar file...')
+    tar.extractall(extracted_data_dir)
+    tar.close()
+    logger.info('Tar file extracted successfully.')
+
+
 def download_and_extract_data():
     """
-    Downloads and extracts the data from the given url to the target directory.
+    Downloads and extracts the data from the given URL to the target directory.
+    If the tar file already exists and was downloaded within the last 29 days, it will be extracted without re-downloading.
 
     :return: None
     :rtype: None
     """
     from config import GSOY_DATA_DIR
+    from logging_config import logger
+    from config import GSOY_DOWNLOAD_URL
+
     os.makedirs(GSOY_DATA_DIR, exist_ok=True)
 
     file_name = os.path.join(GSOY_DATA_DIR, 'gsoy-latest.tar.gz')
+
+    if os.path.exists(file_name):
+        modified_time = datetime.fromtimestamp(os.path.getmtime(file_name))
+        current_time = datetime.now()
+        time_diff = current_time - modified_time
+
+        if time_diff <= timedelta(days=15):
+            logger.info('The tar file already exists and was downloaded within the last 15 days. Skipping download.')
+            if os.path.exists(os.path.join(GSOY_DATA_DIR, 'extracted_data')):
+                logger.info('Extracted data already exists. Skipping extraction.')
+            else:
+                extract_tar_file(file_name)
+            return
+        else:
+            logger.info('The tar file already exists but was downloaded more than 15 days ago. Downloading again.')
+            os.remove(file_name)
+
+    logger.info('Downloading tar file...')
     urllib.request.urlretrieve(GSOY_DOWNLOAD_URL, file_name)
+    logger.info('Tar file downloaded successfully.')
 
-    tar = tarfile.open(file_name, 'r:gz')
-    tar.extractall(GSOY_DATA_DIR)
-    tar.close()
+    extract_tar_file(file_name)
 
-    os.remove(file_name)
+
+def remove_extracted_data():
+    """
+    Removes the previously extracted data directory from the target directory.
+
+    :return: None
+    :rtype: None
+    """
+    from config import GSOY_DATA_DIR
+    from logging_config import logger
+
+    extracted_data_dir = os.path.join(GSOY_DATA_DIR, 'extracted_data')
+
+    if os.path.exists(extracted_data_dir):
+        shutil.rmtree(extracted_data_dir)
+        logger.info('Extracted data removed successfully.')
+    else:
+        logger.info('No extracted data found.')
